102-interpolation.c

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -20,7 +20,8 @@ int interpolation_search(int *array, size_t size, int value)
 
 	while (low < high + 1)
 	{
-		pros = low + ((high - low) / (size_t)(array[high] - array[low])) * (size_t)(value - array[low]);
+		pros = low + (((double)(high - low) / (array[high] - array[low])) *
+				(value - array[low]));
 		if (pros > size - 1)
 		{
 			printf("Value checked array[%ld] is out of range\n", pros);
